+++ tests/scheduler/instrument_coordinator/components/test_qblox.py
@@ -31,7 +31,6 @@ from qcodes.instrument import Instrument, InstrumentChannel, InstrumentModule
 from quantify_core.data.handling import set_datadir  # pylint: disable=no-name-in-module
 
 from quantify_scheduler.compilation import qcompile
-from quantify_scheduler.device_under_test.mock_setup import set_standard_params_transmon
 from quantify_scheduler.device_under_test.transmon_element import BasicTransmonElement
 from quantify_scheduler.instrument_coordinator.components import qblox
 
@@ -550,7 +549,7 @@ def test_prepare_cluster_rf(
 
 def test_prepare_rf(
     close_all_instruments,
-    mock_setup_basic_transmon,
+    mock_setup_basic_transmon_with_standard_params,
     schedule_with_measurement_q2,
     load_example_qblox_hardware_config,
     make_qcm_rf,
@@ -560,11 +559,11 @@ def test_prepare_rf(
     qcm: qblox.QCMRFComponent = make_qcm_rf("qcm_rf0", "1234")
     qrm: qblox.QRMRFComponent = make_qrm_rf("qrm_rf0", "1234")
 
-    set_standard_params_transmon(mock_setup_basic_transmon)
-    mock_setup_basic_transmon["q2"].clock_freqs.readout(7.5e9)
-    mock_setup_basic_transmon["q2"].clock_freqs.f01(6.03e9)
+    mock_setup = mock_setup_basic_transmon_with_standard_params
+    mock_setup["q2"].clock_freqs.readout(7.5e9)
+    mock_setup["q2"].clock_freqs.f01(6.03e9)
 
-    device_config = mock_setup_basic_transmon["quantum_device"].generate_device_config()
+    device_config = mock_setup["quantum_device"].generate_device_config()
     # Act
     with tempfile.TemporaryDirectory() as tmp_dir:
         set_datadir(tmp_dir)
@@ -729,7 +728,7 @@ def test_retrieve_acquisition_qcm_rf(close_all_instruments, make_qcm_rf):
 
 
 def test_retrieve_acquisition_qrm_rf(
-    mock_setup_basic_transmon,
+    mock_setup_basic_transmon_with_standard_params,
     schedule_with_measurement_q2,
     load_example_qblox_hardware_config,
     make_qrm_rf,
@@ -737,9 +736,9 @@ def test_retrieve_acquisition_qrm_rf(
     # Arrange
     qrm_rf: qblox.QRMRFComponent = make_qrm_rf("qrm_rf0", "1234")
 
-    set_standard_params_transmon(mock_setup_basic_transmon)
-    mock_setup_basic_transmon["q2"].clock_freqs.readout(7.3e9)
-    device_config = mock_setup_basic_transmon["quantum_device"].generate_device_config()
+    mock_setup = mock_setup_basic_transmon_with_standard_params
+    mock_setup["q2"].clock_freqs.readout(7.3e9)
+    device_config = mock_setup["quantum_device"].generate_device_config()
 
     # Act
     with tempfile.TemporaryDirectory() as tmp_dir:
@@ -832,7 +831,7 @@ def test_start_qcm_qrm(
 
 
 def test_start_qcm_qrm_rf(
-    mock_setup_basic_transmon,
+    mock_setup_basic_transmon_with_standard_params,
     schedule_with_measurement_q2,
     load_example_qblox_hardware_config,
     make_qcm_rf,
@@ -842,11 +841,11 @@ def test_start_qcm_qrm_rf(
     qcm_rf: qblox.QCMRFComponent = make_qcm_rf("qcm_rf0", "1234")
     qrm_rf: qblox.QRMRFComponent = make_qrm_rf("qrm_rf0", "1234")
 
-    set_standard_params_transmon(mock_setup_basic_transmon)
-    mock_setup_basic_transmon["q2"].clock_freqs.readout(7.3e9)
-    mock_setup_basic_transmon["q2"].clock_freqs.f01(6.03e9)
+    mock_setup = mock_setup_basic_transmon_with_standard_params
+    mock_setup["q2"].clock_freqs.readout(7.3e9)
+    mock_setup["q2"].clock_freqs.f01(6.03e9)
 
-    device_config = mock_setup_basic_transmon["quantum_device"].generate_device_config()
+    device_config = mock_setup["quantum_device"].generate_device_config()
 
     # Act
     with tempfile.TemporaryDirectory() as tmp_dir: